requirements.txt
# Required libraries 
torch
torchvision
pandas
matplotlib
scikit-learn
requests
Pillow
beautifulsoup4